refactor: move TextField component to forms styles
