--- src/components/TextField/index.tsx
@@ -1,9 +1,0 @@
-import { TextField, Theme, withStyles } from '@material-ui/core';
-
-export const StyledTextField = withStyles((theme: Theme) => ({
-  root: {
-    height: '7rem',
-  },
-}))(TextField);
-
-export default StyledTextField;
